Add GuessForm component
This component does a few things that can be refactored
into separate components in future iterations.

What it does:
 - Fetch https://dog.ceo API to get a random photo URL
of a breed of dog
 - Render the form to input and submit the guess
 - Extract the breed name from the URL and set that as
the correct answer
 - Display the requested photo
 - On value change of the input field, set the value
as the current guess value
 - Upon submit, compare the answer and the guess
   - If correct, display the end game view showing
     the answer, number of guesses, and the guesses
     history. There is also a Restart button to reset
     the game.
   - If incorrect, display the guesses history with the
     guessed values from most recent to oldest and
     increment the guess count
- Render a Give Up button to allow users to end the game
and view the result view immediately

We also need to configure NextJS to accept the API domain
as host of the images.